WIP: Add basic spectrum CLI
